chore: 누락된 호출 추가 및 코드 정리
