+++ BoxOffice/Presentation/View/BoxOfficeCollectionViewCell.swift
@@ -1,7 +1,7 @@
 import UIKit
 
 final class BoxOfficeCollectionViewCell: UICollectionViewCell {
-  private var boxOfficeListItem: DailyBoxOffice.ListItem?
+  //  private var boxOfficeListItem: DailyBoxOffice.ListItem?
   
   private var rankLabel: UILabel = {
     let label = UILabel()
@@ -34,7 +34,7 @@ final class BoxOfficeCollectionViewCell: UICollectionViewCell {
   
   private lazy var movieInfoStackView: UIStackView = {
     let stack = UIStackView(arrangedSubviews: [
-    movieNameLabel, audienceLabel
+      movieNameLabel, audienceLabel
     ])
     stack.axis = .horizontal
     stack.spacing = 1
@@ -48,24 +48,36 @@ final class BoxOfficeCollectionViewCell: UICollectionViewCell {
     return imageView
   }()
   
+  init() {
+    super.init(frame: .zero)
+    setLayout()
+  }
+  
+  @available(*, unavailable)
+  required init(coder: NSCoder) {
+    fatalError("init(coder:) has not been implemented")
+  }
+  
   private func setLayout() {
-    self.contentView.addSubview(rankInfoStackView)
-    self.contentView.addSubview(movieInfoStackView)
-    self.contentView.addSubview(chevron)
-    
     self.rankInfoStackView.translatesAutoresizingMaskIntoConstraints = false
     self.movieInfoStackView.translatesAutoresizingMaskIntoConstraints = false
     self.chevron.translatesAutoresizingMaskIntoConstraints = false
-    NSLayoutConstraint.activate([
-      rankInfoStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 20),
-      rankInfoStackView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-      movieInfoStackView.leadingAnchor.constraint(equalTo: rankInfoStackView.trailingAnchor, constant: 15),
-      movieInfoStackView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-      chevron.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -10),
-      chevron.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-      chevron.leadingAnchor.constraint(greaterThanOrEqualTo: movieInfoStackView.trailingAnchor, constant: 0)
-      
-    ])
+    
+    self.contentView.addSubview(self.rankInfoStackView)
+    self.contentView.addSubview(self.movieInfoStackView)
+    self.contentView.addSubview(self.chevron)
+    
+    NSLayoutConstraint.activate(
+      [
+        self.rankInfoStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 20),
+        self.rankInfoStackView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
+        self.movieInfoStackView.leadingAnchor.constraint(equalTo: rankInfoStackView.trailingAnchor, constant: 15),
+        self.movieInfoStackView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
+        self.chevron.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -10),
+        self.chevron.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
+        self.chevron.leadingAnchor.constraint(greaterThanOrEqualTo: movieInfoStackView.trailingAnchor, constant: 0),
+      ]
+    )
   }
   
   func configure(with item: DailyBoxOffice.ListItem) {
@@ -79,60 +91,5 @@ final class BoxOfficeCollectionViewCell: UICollectionViewCell {
     
     self.movieNameLabel.text = item.title
     self.audienceLabel.text = "오늘 \(item.todayAudienceCount) / 총 \(item.totalAudienceCount)"
-  }
-}
-  
-
-final class RiseOrFallStackView: UIStackView {
-  private let riseImage = UIImage(systemName: "arrowtriangle.down.fill")
-  private let fallImage = UIImage(systemName: "arrowtriangle.up.fill")
-  private let minusImage = UIImage(systemName: "minus")
-  
-  private let riseOrFallImageView = UIImageView()
-  private let riseOrFallLabel = UILabel()
-  
-  init() {
-    super.init(frame: .zero)
-    setSubViews()
-    setLayout()
-  }
-  
-  @available(*, unavailable)
-  required init(coder: NSCoder) {
-    fatalError("init(coder:) has not been implemented")
-  }
-  
-  private func setSubViews() {
-    self.addArrangedSubview(riseOrFallImageView)
-    self.addArrangedSubview(riseOrFallLabel)
-  }
-
-  private func setLayout() {
-    self.riseOrFallImageView.translatesAutoresizingMaskIntoConstraints = false
-    self.riseOrFallLabel.translatesAutoresizingMaskIntoConstraints = false
-    NSLayoutConstraint.activate(
-      [
-        riseOrFallImageView.heightAnchor.constraint(equalToConstant: 10),
-        riseOrFallLabel.heightAnchor.constraint(equalTo: riseOrFallImageView.heightAnchor)
-      ]
-    )
-  }
-  
-  func configure(with movieStatus: MovieStatus) {
-    switch movieStatus {
-    case .newMovie:
-      riseOrFallImageView.image = nil
-      riseOrFallLabel.text = "신작"
-      riseOrFallLabel.textColor = .systemPink
-    case let .oldMovie(.risen(number)):
-      riseOrFallImageView.image = riseImage
-      riseOrFallLabel.text = String(number)
-    case let .oldMovie(.fallen(number)):
-      riseOrFallImageView.image = fallImage
-      riseOrFallLabel.text = String(number)
-    case .oldMovie(.maintained):
-      riseOrFallImageView.image = minusImage
-      riseOrFallLabel.text = ""
-    }
   }
 }
